Rename and move config.py and configmodel.py to their own module
