+++ cddagl/launcher.py
@@ -23,7 +23,7 @@ else:
     sys.path.append(basedir)
 
 from cddagl.i18n import load_gettext_no_locale, proxy_gettext as _
-from cddagl.config import init_config, get_config_value, config_true
+from cddagl.sql.functions import init_config, get_config_value, config_true
 from cddagl.ui import start_ui, ui_exception
 
 from cddagl.win32 import get_ui_locale, SingleInstance, write_named_pipe
