+++ cddagl/ui.py
@@ -60,7 +60,7 @@ from PyQt5.QtWidgets import (
 from PyQt5.QtNetwork import QNetworkAccessManager, QNetworkRequest
 
 from cddagl.i18n import load_gettext_locale, proxy_gettext as _, proxy_ngettext as ngettext
-from cddagl.config import (
+from cddagl.sql.functions import (
     get_config_value, set_config_value, new_version, get_build_from_sha256,
     new_build, config_true)
 from cddagl.win32 import (
